+++ backend/canisters/notifications_index/impl/src/jobs/sync_notifications_canisters.rs
@@ -15,7 +15,7 @@ pub(crate) fn start_job_if_required(runtime_state: &RuntimeState) -> bool {
         && (!runtime_state.data.notifications_index_event_sync_queue.is_empty()
             || runtime_state.data.notifications_index_event_sync_queue.sync_in_progress())
     {
-        let timer_id = ic_cdk_timers::set_timer_interval(Duration::default(), run);
+        let timer_id = ic_cdk_timers::set_timer_interval(Duration::ZERO, run);
         TIMER_ID.with(|t| t.set(Some(timer_id)));
         trace!("'sync_notifications_canisters' job started");
         true