+++ backend/canisters/group/api/src/queries/mod.rs
@@ -7,7 +7,6 @@ pub mod events;
 pub mod events_by_index;
 pub mod events_window;
 pub mod http_request;
-pub mod invite_code;
 pub mod local_user_index;
 pub mod messages_by_message_index;
 pub mod public_summary;