+++ backend/integration_tests/src/notification_tests.rs
@@ -1,5 +1,6 @@
 use crate::env::ENV;
 use crate::rng::random_string;
+use crate::utils::tick_many;
 use crate::{client, CanisterIds, TestEnv, User};
 use candid::Principal;
 use ic_test_state_machine_client::StateMachine;
@@ -155,6 +156,8 @@ fn latest_notification_index(env: &StateMachine, notifications_canister_id: Prin
 fn init_test_data(env: &mut StateMachine, canister_ids: &CanisterIds) -> TestData {
     let user1 = client::user_index::happy_path::register_user(env, canister_ids.user_index);
     let user2 = client::user_index::happy_path::register_user(env, canister_ids.user_index);
+
+    tick_many(env, 5);
 
     client::notifications_index::push_subscription(
         env,