+++ frontend/app/src/components/home/ChatList.svelte
@@ -141,6 +141,7 @@
         on:showHomePage
         on:logout
         on:whatsHot
+        on:halloffame
         on:showFaq
         {user}
         on:profile