+++ frontend/app/src/components/home/Home.svelte
@@ -59,6 +59,7 @@
     import { querystring } from "routes";
     import { eventListScrollTop } from "../../stores/scrollPos";
     import GateCheckFailed from "./groupdetails/GateCheckFailed.svelte";
+    import HallOfFame from "./HallOfFame.svelte";
 
     const client = getContext<OpenChat>("client");
     const user = client.user;
@@ -92,6 +93,7 @@
         NewGroup,
         Wallet,
         GateCheckFailed,
+        HallOfFame,
     }
 
     let faqQuestion: Questions | undefined = undefined;
@@ -183,13 +185,6 @@
         // if this is an unknown chat let's preview it
         if (chat === undefined) {
             if (!(await createDirectChat(chatId))) {
-                const code = $querystring.get("code");
-                if (code) {
-                    client.groupInvite = {
-                        chatId,
-                        code,
-                    };
-                }
                 if (!(await client.previewChat(chatId))) {
                     page.replace("/");
                     return;
@@ -292,6 +287,12 @@
                 if (wallet !== null) {
                     modal = ModalType.Wallet;
                     page.replace(removeQueryStringParam("wallet"));
+                }
+
+                const hof = $querystring.get("hof");
+                if (hof !== null) {
+                    modal = ModalType.HallOfFame;
+                    page.replace(removeQueryStringParam("hof"));
                 }
             }
         }
@@ -907,6 +908,7 @@
             on:searchEntered={performSearch}
             on:chatWith={chatWith}
             on:whatsHot={() => whatsHot(true)}
+            on:halloffame={() => (modal = ModalType.HallOfFame)}
             on:newGroup={newGroup}
             on:profile={showProfile}
             on:logout={() => client.logout()}
@@ -1022,6 +1024,8 @@
             <NewGroup on:upgrade={upgrade} {candidateGroup} on:close={closeModal} />
         {:else if modal === ModalType.Wallet}
             <AccountsModal on:close={closeModal} />
+        {:else if modal === ModalType.HallOfFame}
+            <HallOfFame on:close={closeModal} />
         {/if}
     </Overlay>
 {/if}