+++ frontend/openchat-agent/src/services/groupIndex/candid/idl.js
@@ -90,16 +90,6 @@ export const idlFactory = ({ IDL }) => {
     'frozen_by' : UserId,
     'reason' : IDL.Opt(IDL.Text),
   });
-  const MessageReport = IDL.Record({
-    'notes' : IDL.Opt(IDL.Text),
-    'timestamp' : TimestampMillis,
-    'reported_by' : UserId,
-    'reason_code' : IDL.Nat32,
-  });
-  const ReportedMessage = IDL.Record({
-    'count' : IDL.Nat32,
-    'reports' : IDL.Vec(MessageReport),
-  });
   const GiphyImageVariant = IDL.Record({
     'url' : IDL.Text,
     'height' : IDL.Nat32,
@@ -369,7 +359,6 @@ export const idlFactory = ({ IDL }) => {
     'reminder_id' : IDL.Nat64,
   });
   const MessageContent = IDL.Variant({
-    'ReportedMessage' : ReportedMessage,
     'Giphy' : GiphyContent,
     'File' : FileContent,
     'Poll' : PollContent,