+++ frontend/openchat-shared/src/domain/chat/chat.ts
@@ -1484,6 +1484,7 @@ export type JoinGroupResponse =
     | GroupChatSummary
     | { kind: "blocked" }
     | { kind: "group_not_found" }
+    | { kind: "not_invited" }
     | { kind: "group_not_public" }
     | { kind: "already_in_group" }
     | { kind: "not_super_admin" }
@@ -1620,36 +1621,6 @@ export type RegisterPollVoteResponse =
     | "chat_frozen"
     | "polls_not_valid_for_direct_chats";
 
-export type InviteCodeResponse = InviteCodeSuccess | NotAuthorised;
-
-export type InviteCodeSuccess = {
-    kind: "success";
-    code?: string;
-};
-
-export type EnableInviteCodeResponse =
-    | EnableInviteCodeSuccess
-    | NotAuthorised
-    | UserSuspended
-    | ChatFrozen;
-
-export type EnableInviteCodeSuccess = {
-    kind: "success";
-    code: string;
-};
-
-export type DisableInviteCodeResponse =
-    | "not_authorised"
-    | "user_suspended"
-    | "chat_frozen"
-    | "success";
-
-export type ResetInviteCodeResponse =
-    | ResetInviteCodeSuccess
-    | NotAuthorised
-    | UserSuspended
-    | ChatFrozen;
-
 export type ThreadPreviewsResponse = CallerNotInGroup | ThreadPreviewsSuccess;
 
 export type ThreadPreviewsSuccess = {
@@ -1662,11 +1633,6 @@ export type ThreadPreview = {
     latestReplies: EventWrapper<Message>[];
     totalReplies: number;
     rootMessage: EventWrapper<Message>;
-};
-
-export type ResetInviteCodeSuccess = {
-    kind: "success";
-    code: string;
 };
 
 export type MessageAction = "emoji" | "file" | undefined;