+++ frontend/openchat-worker/src/worker.ts
@@ -943,13 +943,6 @@ self.addEventListener("message", (msg: MessageEvent<WorkerRequest>) => {
                     .catch(sendError(correlationId));
                 break;
 
-            case "groupInvite":
-                agent.groupInvite = payload.value;
-                sendResponse(correlationId, {
-                    response: undefined,
-                });
-                break;
-
             case "searchGroupChat":
                 agent
                     .searchGroupChat(
@@ -1072,39 +1065,6 @@ self.addEventListener("message", (msg: MessageEvent<WorkerRequest>) => {
                         payload.messageIndexes,
                         payload.latestClientEventIndex
                     )
-                    .then((response) =>
-                        sendResponse(correlationId, {
-                            response,
-                        })
-                    )
-                    .catch(sendError(correlationId));
-                break;
-
-            case "getInviteCode":
-                agent
-                    .getInviteCode(payload.chatId)
-                    .then((response) =>
-                        sendResponse(correlationId, {
-                            response,
-                        })
-                    )
-                    .catch(sendError(correlationId));
-                break;
-
-            case "enableInviteCode":
-                agent
-                    .enableInviteCode(payload.chatId)
-                    .then((response) =>
-                        sendResponse(correlationId, {
-                            response,
-                        })
-                    )
-                    .catch(sendError(correlationId));
-                break;
-
-            case "disableInviteCode":
-                agent
-                    .disableInviteCode(payload.chatId)
                     .then((response) =>
                         sendResponse(correlationId, {
                             response,
@@ -1333,6 +1293,17 @@ self.addEventListener("message", (msg: MessageEvent<WorkerRequest>) => {
             case "cancelMessageReminder":
                 agent
                     .cancelMessageReminder(payload.reminderId)
+                    .then((response) =>
+                        sendResponse(correlationId, {
+                            response,
+                        })
+                    )
+                    .catch(sendError(correlationId));
+                break;
+
+            case "getReferralLeaderboard":
+                agent
+                    .getReferralLeaderboard(payload.args)
                     .then((response) =>
                         sendResponse(correlationId, {
                             response,