+++ backend/canisters/user_index/CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Added `platform_moderators_group` query endpoint ([#3495](https://github.com/open-ic/open-chat/pull/3495))
 - Join users who register with relevant code to Bitcoin Miami group ([#3501](https://github.com/open-ic/open-chat/pull/3501))
 
+### Changed
+
+- Reduce a few timer job intervals ([#3515](https://github.com/open-ic/open-chat/pull/3515))
+
 ## [[2.0.668](https://github.com/open-ic/open-chat/releases/tag/v2.0.668-user_index)] - 2023-04-19
 
 ### Changed