+++ backend/canisters/user_index/CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+### Changed
+
+- Pass OpenChat bot messages in user canister init args ([#3517](https://github.com/open-ic/open-chat/pull/3517))
+
 ## [[2.0.671](https://github.com/open-ic/open-chat/releases/tag/v2.0.671-user_index)] - 2023-04-28
 
 ### Added